requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astroid==2.5.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
black==21.9b0
bleach==3.3.0
cffi==1.14.5
click==8.0.3
cvxopt==1.2.6
cvxpy==1.1.15
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
ecos==2.0.7.post1
entrypoints==0.3
flake8==3.9.0
guppy3==3.1.1
gurobipy==9.1.2
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mistune==0.8.4
Mosek==9.3.6
mypy==0.812
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.21.3
opt-einsum==3.3.0
osqp==0.6.2.post0
packaging==20.9
pandocfilters==1.4.3
parameterized==0.8.1
parso==0.8.2
pathspec==0.9.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.8.1
pylint==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pyzmq==22.0.3
qdldl==0.1.5.post0
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.10.23
scikit-learn==0.24.2
scipy==1.7.0
scs==2.1.4
Send2Trash==1.5.0
simplejson==3.17.3
six==1.15.0
sklearn==0.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tomli==1.2.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
txt2tags==3.7
typed-ast==1.4.3
typing-extensions==3.10.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
